Split api implementation
API handler is in apihandler.c
API callbacks (called by handler) are in api.c
Renamed /token to /register
